refactor(core): moves incremental hydration codebase to better locations
This eliminates the extra incremental.ts codepath and moves its functions to appropriate locations.
