+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -1593,9 +1593,6 @@
     "name": "init_image_performance_warning"
   },
   {
-    "name": "init_incremental"
-  },
-  {
     "name": "init_inert_body"
   },
   {
